+++ lib/ui/settings/SettingsHomeContentPanel.dart
@@ -20,6 +20,7 @@ import 'package:flutter/cupertino.dart';
 import 'package:illinois/service/Analytics.dart';
 import 'package:illinois/ui/athletics/AthleticsTeamsWidget.dart';
 import 'package:illinois/ui/home/HomePanel.dart';
+import 'package:illinois/ui/settings/SettingsAppointmentsContentWidget.dart';
 import 'package:illinois/ui/settings/SettingsCalendarContentWidget.dart';
 import 'package:illinois/ui/settings/SettingsFoodFiltersContentWidget.dart';
 import 'package:illinois/ui/settings/SettingsInterestsContentWidget.dart';
@@ -192,6 +193,8 @@ class _SettingsHomeContentPanelState extends State<SettingsHomeContentPanel> {
         return AthleticsTeamsWidget();
       case SettingsContent.calendar:
         return SettingsCalendarContentWidget();
+      case SettingsContent.appointments:
+        return SettingsAppointmentsContentWidget();
       case SettingsContent.favorites:
         return Container();
     }
@@ -217,13 +220,15 @@ class _SettingsHomeContentPanelState extends State<SettingsHomeContentPanel> {
         return Localization().getStringEx('panel.settings.home.settings.sections.sports.label', 'My Sports Teams');
       case SettingsContent.calendar:
         return Localization().getStringEx('panel.settings.home.settings.sections.calendar.label', 'My Calendar Settings');
+      case SettingsContent.appointments:
+        return Localization().getStringEx('panel.settings.home.settings.sections.appointments.label', 'MyMcKinley Appointments');
       case SettingsContent.favorites:
         return Localization().getStringEx('panel.settings.home.settings.sections.favorites.label', 'My Favorites');
     }
   }
 }
 
-enum SettingsContent { sections, interests, food_filters, sports, calendar, favorites }
+enum SettingsContent { sections, interests, food_filters, sports, calendar, appointments, favorites }
 
 class _DebugContainer extends StatefulWidget implements PreferredSizeWidget {
   final Widget _child;